requirements.txt
streamlit
python-dotenv
unstructured
sentence-transformers
chromadb
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-text-splitters==0.0.2
langchain-groq==0.1.3
langchain-huggingface==0.1.2
groq
huggingface-hub
torch
